[move-only] rpc: move blockchain rpcs to a new mod
If we do all rpcs inside `server.rs` it'll grow and become a huge file,
this commit creates a blockchain module where we define all
blockchain-related rpcs as defined by core[1] docs.

[1]: https://bitcoincore.org/en/doc/28.0.0/rpc
