+++ florestad/src/json_rpc/mod.rs
@@ -1,2 +1,3 @@
+pub mod blockchain;
 pub mod res;
 pub mod server;
